Review feedback
Remove commented code. Add Kubeconfig from env
